Add delete question options and refactor payment handler
Added new boolean properties 'deleteQuestion' and 'deleteQuestionButtons' to conversation steps  to provide more control over the conversation flow.

Refactored handleSuccessfulPayment method by renaming '$payment' parameter to '$data' for better readability. Modified cleanup closure in handleConversations to consider new properties. Now we can decide to delete a question or its buttons separately after a step.

These changes

+++ src/Concerns/HandlesConversations.php
@@ -115,7 +115,11 @@ trait HandlesConversations
 
         $variations[self::CLEANUP_CLOSURE_NAME] = function () use ($messageId) {
             /** @var Conversation $this */
-            if (isset($this->chat)) {
+            if (isset($this->chat) && $this->chat->step->deleteQuestion) {
+                $this->chat->deleteMessage($messageId)->send();
+            }
+
+            if (isset($this->chat) && $this->chat->step->deleteQuestionButtons) {
                 $this->chat->deleteKeyboard($messageId)->send();
             }
         };
